feat: prefechQuery 추상화
