+++ src/app/(public)/signup/page.tsx
@@ -1,8 +1,6 @@
-import { getMe } from '@/apis/auth.api';
-import { QUERY_KEY_ME } from '@/constants/auth.const';
 import SignUpContainer from '@/containers/auth/SignUpContainer';
-import { getCookie } from '@/utils/auth-server.util';
-import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
+import { getCookie, prefetchMe } from '@/utils/auth-server.util';
+import { HydrationBoundary } from '@tanstack/react-query';
 import { redirect } from 'next/navigation';
 import { Suspense } from 'react';
 
@@ -10,14 +8,7 @@ export default async function SignUpPage() {
   const token = await getCookie('dudemeet-token');
   if (!token) return <SignUpContainer />;
 
-  const queryClient = new QueryClient();
-
-  await queryClient.prefetchQuery({
-    queryKey: [QUERY_KEY_ME],
-    queryFn: () => getMe(),
-  });
-  const me = await queryClient.getQueryData([QUERY_KEY_ME]);
-  const dehydratedState = dehydrate(queryClient);
+  const { me, dehydratedState } = await prefetchMe();
 
   if (me) redirect('/');
